refactor: move ModalStack inside App container
